+++ frdrpc/gen_protos.sh
@@ -7,17 +7,25 @@ function generate() {
   echo "Generating faraday gRPC server protos"
 
   # Generate the protos.
-  protoc --go_out=plugins=grpc,paths=source_relative:. \
+  protoc -I/usr/local/include -I. \
+    --go_out . --go_opt paths=source_relative \
+    --go-grpc_out . --go-grpc_opt paths=source_relative \
     faraday.proto
 
   # Generate the REST reverse proxy.
-  protoc \
-    --grpc-gateway_out=logtostderr=true,paths=source_relative,grpc_api_configuration=rest-annotations.yaml:. \
+  protoc -I/usr/local/include -I. \
+    --grpc-gateway_out . \
+    --grpc-gateway_opt logtostderr=true \
+    --grpc-gateway_opt paths=source_relative \
+    --grpc-gateway_opt grpc_api_configuration=faraday.yaml \
     faraday.proto
 
   # Finally, generate the swagger file which describes the REST API in detail.
-  protoc \
-    --swagger_out=logtostderr=true,grpc_api_configuration=rest-annotations.yaml:. \
+  protoc -I/usr/local/include -I. \
+    --openapiv2_out . \
+    --openapiv2_opt logtostderr=true \
+    --openapiv2_opt grpc_api_configuration=faraday.yaml \
+    --openapiv2_opt json_names_for_fields=false \
     faraday.proto
 }